requirements.txt
Django
djangorestframework==3.14.0
pillow==10.2.0
django-phonenumber-field==7.3.0
phonenumbers